Api manoj ll

+++ brahmi_backend/controller.py
@@ -1,7 +1,8 @@
 import os
-import json
-from flask import Flask, redirect, url_for, request, jsonify, Response, abort
-
+from flask import Flask, redirect, url_for, request, jsonify, Response, abort, json
+import filetype
+import services
+from util import make_response
 app = Flask(__name__)
 
 
@@ -11,22 +12,21 @@ def translateLetters():
         image = request.files["image"]
         image_name = image.filename
         image.save(os.path.join(os.getcwd(), image_name))
-        # todo: pre processing part
-        # todo :  predictions
-        # todo : get result and add to response
-        response = ['a', 'b', 'c', 'd']
-        letter = []
-        count = 1
-
-        for x in response:
-            x = {count: x}
-            letter.append(x)
-            count = count + 1
-        os.remove(image_name)
-
-        return Response(json.dumps(letter), status=200, mimetype='application/json')
-    except FileNotFoundError:
-        abort(404)
+        if filetype.is_image(image_name):
+            classify_letters = services.classify_letters(image_name)
+            result = {'letters': classify_letters,'suggestions':{'c':1}}
+            response = make_response(result,True,200)
+            # todo: pre processing part
+            # todo :  predictions
+            # todo : get result and add to response
+            os.remove(image_name)
+            return Response(response=response, status=200, mimetype='application/json')
+        else:
+            response = make_response('The file is NOT an Image', False, 200)
+            return Response(response=response, status=200, mimetype='application/json')
+    except:
+        response = make_response('The file is NOT FOUND', False, 404)
+        return Response(response=response, status=404, mimetype='application/json')
 
 
 # @app.route("/api/getWords", methods=["POST"])
